refactor|test: Add tests for harvest periods and clean code
