+++ contracts/IAOUpgradeProxy.sol
@@ -13,7 +13,7 @@ pragma solidity 0.8.6;
 
 import "@openzeppelin/contracts/proxy/transparent/TransparentUpgradeableProxy.sol";
 
-contract IFOUpgradeProxy is TransparentUpgradeableProxy {
+contract IAOUpgradeProxy is TransparentUpgradeableProxy {
 
     constructor(address admin, address logic, bytes memory data) TransparentUpgradeableProxy(logic, admin, data) {
